First set computation (needs verification with our manual submission)

+++ src/parser/set.c
@@ -1,4 +1,5 @@
 #include "set.h"
+#include <stdlib.h>
 
 token_set *createTokenSet()
 {
@@ -9,15 +10,15 @@ token_set *createTokenSet()
 
 void addToken(token_set *ts, TOKEN tok)
 {
-    ts->set |= (1 << tok);
+    ts->set |= ((long long)1 << tok);
 }
 
 void removeToken(token_set *ts, TOKEN tok)
 {
-    ts->set &= ~(1 << tok);
+    ts->set &= ~((long long)1 << tok);
 }
 
 int isMember(token_set *ts, TOKEN tok)
 {
-    return (ts->set & (1 << tok)) != 0;
+    return (ts->set & ((long long)1 << tok)) != (long long)0;
 }
